Added lots of new features with even more ideas on the way.
Docking at the bottom and top of the screen has been implemented, correct use of EMWH has been implemented, as has text drawing with pango. The text and colours of the bar can now be updated as well. New config options have been added, with a fewer amount being removed.

In the future, I hope to add some more documentation, add a more object based approach to rendering as well as deciding upon a way to pass data to barney- at the moment it is just through STDIN.